music-store-recommendations/src/test/java/edu/smu/musicstorerecommendations/controller/AlbumRecommendationControllerTest.java
package edu.smu.musicstorerecommendations.controller;


import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.databind.json.JsonMapper;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import edu.smu.musicstorerecommendations.dto.AlbumRecommendation;
import edu.smu.musicstorerecommendations.service.ServiceLayer;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import java.util.Arrays;
import java.util.HashMap;
import java.util.List;
import java.util.Optional;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(AlbumRecommendationController.class)
public class AlbumRecommendationControllerTest {

    @MockBean
    ServiceLayer serviceLayer;

    @Autowired
    private MockMvc mockMvc;

    private ObjectMapper mapper = JsonMapper.builder()
            .addModule(new JavaTimeModule())
            .build();

    private List<AlbumRecommendation> expectedAlbumRecommendationList;
    private AlbumRecommendation expectedAlbumRecommendation;
    private AlbumRecommendation inputtedAlbumRecommendation;
    private String expectedJson;
    private String inputtedJson;

    @Before
    public void setUpAndMocking() {
        expectedAlbumRecommendationList = Arrays.asList(
                new AlbumRecommendation(),
                new AlbumRecommendation()
        );
        expectedAlbumRecommendation = new AlbumRecommendation();
        inputtedAlbumRecommendation = new AlbumRecommendation();

        when(serviceLayer.getAllAlbumRecommendations()).thenReturn(expectedAlbumRecommendationList);
        when(serviceLayer.getAlbumRecommendationById(2)).thenReturn(Optional.of(expectedAlbumRecommendation));
        when(serviceLayer.createAlbumRecommendation(inputtedAlbumRecommendation)).thenReturn(expectedAlbumRecommendation);
    }

    /* ============================= TESTING GET ROUTES ============================= */
    /* --------------------------------- HAPPY PATHS -------------------------------- */
    @Test
    public void shouldReturnListOfAllAlbumRecommendationsAndStatus200() throws Exception {
        expectedJson = mapper.writeValueAsString(expectedAlbumRecommendationList);

        mockMvc.perform(get("/albumRecommendation"))
                .andDo(print())
                .andExpect(content().json(expectedJson))
                .andExpect(status().isOk());
    }

    @Test
    public void shouldReturnAlbumRecommendationByIdAndStatus200() throws Exception {
        expectedJson = mapper.writeValueAsString(expectedAlbumRecommendation);

        mockMvc.perform(get("/albumRecommendation/2"))
                .andDo(print())
                .andExpect(content().json(expectedJson))
                .andExpect(status().isOk());
    }

    /* ---------------------------------- SAD PATHS --------------------------------- */
    @Test
    public void shouldReturn404StatusCodeIfAlbumRecommendationIdDoesNotExist() throws Exception {
        mockMvc.perform(get("/albumRecommendation/10023"))
                .andDo(print())
                .andExpect(status().isNotFound());
    }

    /* ============================= TESTING POST ROUTES ============================= */
    /* --------------------------------- HAPPY PATHS -------------------------------- */
    @Test
    public void shouldReturnAlbumRecommendationOnPostRequestAndStatus201() throws Exception {
        expectedJson = mapper.writeValueAsString(expectedAlbumRecommendation);
        inputtedJson = mapper.writeValueAsString(inputtedAlbumRecommendation);

        mockMvc.perform(post("/albumRecommendation")
                        .content(inputtedJson)
                        .contentType(MediaType.APPLICATION_JSON))
                .andDo(print())
                .andExpect(content().json(expectedJson))
                .andExpect(status().isCreated());
    }

    /* ---------------------------------- SAD PATHS --------------------------------- */
    @Test
    public void ShouldReturnStatus422ForInvalidRequestBodyOnPostRequest() throws Exception {
        HashMap<String, Object> invalidRequestBody = new HashMap();
        invalidRequestBody.put("id", Integer.parseInt("136"));
        invalidRequestBody.put("title", "FakeGameTitle1223");
        invalidRequestBody.put("releaseData", "2022-10-12");

        inputtedJson = mapper.writeValueAsString(invalidRequestBody);

        mockMvc.perform(post("/albumRecommendation")
                        .content(inputtedJson)
                        .contentType(MediaType.APPLICATION_JSON))
                .andDo(print())
                .andExpect(status().isUnprocessableEntity());
    }

    /* ============================= TESTING PUT ROUTES ============================= */
    /* --------------------------------- HAPPY PATHS -------------------------------- */
    @Test
    public void shouldRespondWithStatus204WithValidPutRequest() throws Exception {
        inputtedJson = mapper.writeValueAsString(inputtedAlbumRecommendation);

        mockMvc.perform(put("/albumRecommendation/2")
                        .content(inputtedJson)
                        .contentType(MediaType.APPLICATION_JSON))
                .andDo(print())
                .andExpect(status().isNoContent());
    }
    /* ---------------------------------- SAD PATHS --------------------------------- */
    @Test
    public void shouldReturn422StatusCodeIfAlbumRecommendationIdsDoNotMatch() throws Exception {
        inputtedJson = mapper.writeValueAsString(inputtedAlbumRecommendation);

        mockMvc.perform(put("/albumRecommendation/69421")
                        .content(inputtedJson)
                        .contentType(MediaType.APPLICATION_JSON))
                .andDo(print())
                .andExpect(status().isUnprocessableEntity());
    }

    @Test public void shouldReturn422StatusCodeIfRequestBodyIsInvalidForPutRequest() throws Exception {
        HashMap<String, Object> invalidRequestBody = new HashMap();
        invalidRequestBody.put("id", Integer.parseInt("136"));
        invalidRequestBody.put("title", "FakeGameTitle1223");
        invalidRequestBody.put("releaseDate", "2022-10-12");

        inputtedJson = mapper.writeValueAsString(invalidRequestBody);

        mockMvc.perform(put("/albumRecommendation/136")
                        .content(inputtedJson)
                        .contentType(MediaType.APPLICATION_JSON))
                .andDo(print())
                .andExpect(status().isUnprocessableEntity());
    }

    /* ============================= TESTING DELETE ROUTES ============================= */
    /* --------------------------------- HAPPY PATHS -------------------------------- */
    @Test
    public void shouldRespondWithStatus204WithValidDeleteRequest() throws Exception {
        mockMvc.perform(delete("/albumRecommendation/2"))
                .andDo(print())
                .andExpect(status().isNoContent());
    }

    /* ---------------------------------- SAD PATHS --------------------------------- */
    @Test public void shouldResponseWithStatus404IfGameIdIsNotFoundForDelete() throws Exception {
        mockMvc.perform(delete("/albumRecommendation/1412"))
                .andDo(print())
                .andExpect(status().isNotFound());
    }
}